feat: allow multiple open modals (#7660)
