+++ src/context/ModalProvider/ModalContext.tsx
@@ -1,13 +1,12 @@
 import { createContext } from 'react'
 
 import { MODAL_KEYS } from './constants'
-import type { BaseProps, ModalContext, Modals } from './types'
+import type { BaseProps, Modals, ModalState } from './types'
 
-export const modalContext: ModalContext = MODAL_KEYS.reduce<ModalContext>(
+export const modalContext: ModalState = MODAL_KEYS.reduce<ModalState>(
   (acc, key: keyof Modals) => ({
     ...acc,
-    // blatant cast as we inject the correct props during provider creation in ModalContainer
     [key]: createContext<BaseProps<keyof Modals>>({} as BaseProps<keyof Modals>),
   }),
-  {} as ModalContext,
+  {} as ModalState,
 )
